feat(nodes): rebuild the blockmap for big/limit-removing maps (fixes LoR MAP13+ crash)
The DOOM blockmap stores list offsets as 16-bit, so a blockmap over 65535
shorts can't be addressed -- Legacy of Rust MAP13's is 76134, and every
extended-node map with >32767 line indices also overflowed the signed-short
list entries. Both crashed collision (bogus line pointers).

- blockmap/blockmaplump are now int* (no 16-bit limit).
- P_LoadBlockMap reads the WAD blockmap UNSIGNED into ints (0xFFFF list
  terminator kept as -1), or rebuilds it with P_CreateBlockMap (jff/killough,
  via MBF/Woof) when the lump is missing, degenerate, too big, or -blockmap
  is given.
- Move the P_LoadBlockMap call after vertexes/linedefs load, since a rebuild
  reads lines[]/vertexes[] (it produced a garbage -511x-511 blockmap first).
- P_BlockLinesIterator walks int entries.

Verified: LoR MAP13 (XNOD + a ZNOD recompress) loads, renders and collides;
DOOM2 and LoR MAP01/07 unchanged.

+++ files/p_local.h
@@ -252,8 +252,8 @@ P_RadiusAttack
 // P_SETUP
 //
 extern byte*		rejectmatrix;	// for fast sight rejection
-extern short*		blockmaplump;	// offsets in blockmap are from here
-extern short*		blockmap;
+extern int*		blockmaplump;	// int (not short): built as/loaded into a limit-removing
+extern int*		blockmap;	// int blockmap so big maps + >32767 line indices work
 extern int		bmapwidth;
 extern int		bmapheight;	// in mapblocks
 extern fixed_t		bmaporgx;
